create view model for recommended series screen with effect and interaction listener
