+++ presentation/details/src/main/java/com/giraffe/details/screens/recommended/series/RecommendedInteractionListener.kt
@@ -1,0 +1,5 @@
+package com.giraffe.details.screens.recommended.series
+
+interface RecommendedInteractionListener {
+    fun navigateToSeriesDetailsScreen(seriesId: Int)
+}
