+++ presentation/details/src/main/java/com/giraffe/details/screens/recommended/series/RecommendedSeriesEffect.kt
@@ -1,0 +1,7 @@
+package com.giraffe.details.screens.recommended.series
+
+
+interface RecommendedSeriesEffect {
+    data class Error(val error: Throwable) : RecommendedSeriesEffect
+    data class NavigateToSeriesDetails(val seriesId: Int) : RecommendedSeriesEffect
+}
